+++ src/core/client/events/keyup.ts
@@ -1,7 +1,7 @@
 import * as alt from 'alt-client';
 import * as native from 'natives';
 import { Events_Misc } from '../../shared/enums/events';
-import { toggleInteractionMode } from '../systems/interaction';
+import { toggleInteractionMode, triggerInteraction } from '../systems/interaction';
 import { focusChat, focusLeaderBoard } from '../views/hud/hud';
 
 const DELAY_BETWEEN_PRESSES = 500;
@@ -9,6 +9,7 @@ const keyupBinds = {
     112: handleDebugMessages, // F1
     113: focusLeaderBoard, // F2
     84: focusChat, // T
+    69: triggerInteraction, // E
     18: toggleInteractionMode // alt
 };