+++ src/core/server/commands/commands.ts
@@ -1,4 +1,5 @@
 import './adminChat';
+import './currency';
 import './getVehicle';
 import './noclip';
 import './revive';