+++ src/core/server/extensions/playerPrototypes/sound.ts
@@ -1,0 +1,5 @@
+import { System_Events_Sound } from '../../../shared/enums/system';
+
+export function playFrontendSoundPrototype(audioName: string, ref: string): void {
+    this.emit(System_Events_Sound.PlaySoundFrontend, audioName, ref);
+}